+++ NEWS.md
@@ -1,3 +1,8 @@
+## rtsdata 0.1.2
+
+* Added compression using 'brotli' package.
+* Added full.update flag to the getSymbols function to force full update.
+
 ## rtsdata 0.1.1
 
 Initial release.